refactor: sse 연결 관련 기능 패키지 분리
