+++ backend/turip-app/src/main/java/turip/favorite/stream/controller/dto/response/HeartbeatStreamResponse.java
@@ -1,4 +1,4 @@
-package turip.favorite.controller.dto.response;
+package turip.favorite.stream.controller.dto.response;
 
 import java.time.LocalDateTime;
