+++ backend/turip-app/src/main/java/turip/favorite/stream/service/ActionType.java
@@ -1,4 +1,4 @@
-package turip.favorite.service;
+package turip.favorite.stream.service;
 
 public enum ActionType {
     // FavoritePlace 관련
