+++ backend/turip-app/src/main/java/turip/favorite/stream/service/StreamEventType.java
@@ -1,4 +1,4 @@
-package turip.favorite.service;
+package turip.favorite.stream.service;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
